requirements.txt
alembic==1.14.0
aniso8601==9.0.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
flask-bcrypt
flask-cors
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
Mako==1.3.8
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
outcome==1.3.0.post0
packaging==24.2
psycopg==3.2.3
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
trio==0.27.0
trio-websocket==0.11.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
psycopg2-binary<3.0,>=2.9


